% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_reads}
\alias{all_reads}
\title{all reads sizes table}
\format{An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 36 rows and 12 columns.}
\usage{
all_reads
}
\description{
all reads sizes table
}
\keyword{datasets}
