% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antiviral_cov_tbl}
\alias{antiviral_cov_tbl}
\title{antigenomic viral cov table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 71712 rows and 7 columns.}
\usage{
antiviral_cov_tbl
}
\description{
antigenomic viral cov table
}
\keyword{datasets}
