% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{calc_dinuc_percent}
\alias{calc_dinuc_percent}
\title{Calculate percent of dinucloetide capture per total umi corrected reads in the library}
\usage{
calc_dinuc_percent(filenames, table)
}
\arguments{
\item{filenames}{list of filenames and locations contianing the library information}

\item{table}{table summarized counts for each dinucleotide for each library}
}
\description{
Calculate percent of dinucloetide capture per total umi corrected reads in the library
}
\examples{

calc_dinuc_percent(filenames, obs_freqs)

}
