% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{complete_table}
\alias{complete_table}
\title{Complete coverage table by filling in missing data and adding in dinucleotide information by position}
\usage{
complete_table(table, dinuc_path)
}
\arguments{
\item{x}{table with position and count data}

\item{y}{path to a table with dinucleotide data by position}
}
\value{
table with position, count, and dinucleotide data
}
\description{
Complete coverage table by filling in missing data and adding in dinucleotide information by position
}
\examples{
complete_table(table, dinuc_path)


}
