% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{coverage_table}
\alias{coverage_table}
\title{Load count and posiiton data into table by cell, virus, and time}
\usage{
coverage_table(filenames)
}
\arguments{
\item{x}{list of file names}
}
\value{
table with count data organized by cell, virus, and time
}
\description{
Load count and posiiton data into table by cell, virus, and time
}
\examples{
coverage_table(filenames)


}
