% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endou-package.R
\docType{package}
\name{endoU}
\alias{endoU}
\alias{endoU-package}
\title{endoU: RNA-seq analysis of endoU cleavage sites}
\description{
What the package does (one paragraph).
}
\author{
Rachel Ancar <rachel.ancar@ucdenver.edu>

Jay Hesselberth <jay.hesselberth@gmail.com>
}
