% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{endoU_colors}
\alias{endoU_colors}
\title{Function to extract endoU colors as hex codes}
\format{An object of class \code{character} of length 9.}
\usage{
endoU_colors
}
\description{
Function to extract endoU colors as hex codes
}
\keyword{datasets}
