% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{data}
\name{endoU_palettes}
\alias{endoU_palettes}
\title{Return function to interpolate a endoU color palette}
\format{An object of class \code{list} of length 6.}
\usage{
endoU_palettes
}
\arguments{
\item{palette}{Character name of palette in endoU_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a endoU color palette
}
\keyword{datasets}
