% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{fisher.byrow}
\alias{fisher.byrow}
\title{Calculate significance of enrichment values comparing dinucleotides captured in the
experimental data compared to prevalence of dinucelotide in RNA sequence of interest}
\usage{
fisher.byrow(x)
}
\arguments{
\item{x}{table generated from calculating dinucelotide enrichment}
}
\description{
Calculate significance of enrichment values comparing dinucleotides captured in the
experimental data compared to prevalence of dinucelotide in RNA sequence of interest
}
\examples{
cal_enrichment_sigs(combined_freqs)



function to execute fisher exact test by row of dataframe comapring the odds ratio
of obtaining a speciic dinucleotide in the expected data to the observed data
}
