% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_analysis.R
\name{foldChange}
\alias{foldChange}
\title{Calculate fold change in cleavage activity in the abscence of RNaseL or EndoU actviity}
\usage{
foldChange(spread_tbl, x_col, m_col, x_col2 = NULL)
}
\arguments{
\item{m_col}{Name of mutant cell type or viral mutant of interest}

\item{data}{viral_cov_tbl}

\item{key}{column for comparision, either "cell" or "virus"}
}
\description{
Calculate fold change in cleavage activity in the abscence of RNaseL or EndoU actviity
}
\examples{

foldChange_all(viral_cov_table,
   "RNaseL",
   "cell"
)


Function to calculate log2 fold change comparing mutant/WT
}
