% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{frequencies}
\alias{frequencies}
\title{Calculate dinucleotide frequencies observed from experimntal data}
\usage{
frequencies(x)
}
\arguments{
\item{x}{viral_cov_tbl}
}
\description{
Calculate dinucleotide frequencies observed from experimntal data
}
\examples{

exp_freqs <- calc_freqs(viral_cov_tbl)


function to generate counts of dinucleotides per grouped library
}
