% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{list_names}
\alias{list_names}
\title{Read position and count data into table}
\usage{
list_names(path)
}
\arguments{
\item{x}{path to file names}
}
\description{
Read position and count data into table
}
\examples{
"~/Dropbox (Hesselberth Lab)/Rachel_data/EndoU_project/viral_bg/neg"
filenames <- list_names(x)

}
