% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{load_freqs}
\alias{load_freqs}
\title{Load frequencies from a file}
\usage{
load_freqs(x)
}
\arguments{
\item{x}{tsv file with frequencies}
}
\description{
Load frequencies from a file
}
\examples{
# path to file
load_freqs(x)

}
