% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mhv_freq}
\alias{mhv_freq}
\title{MHV background dinucleotide frequencies table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 3 columns.}
\usage{
mhv_freq
}
\description{
MHV background dinucleotide frequencies table
}
\keyword{datasets}
