% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mRNA_cov_tbl}
\alias{mRNA_cov_tbl}
\title{mRNA cov table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 572583 rows and 7 columns.}
\usage{
mRNA_cov_tbl
}
\description{
mRNA cov table
}
\keyword{datasets}
