% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{no_dinuc_complete}
\alias{no_dinuc_complete}
\title{For generating a position table without dinucleotides}
\usage{
no_dinuc_complete(table)
}
\description{
For generating a position table without dinucleotides
}
\examples{
no_dinuc_complete(table)


}
