% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{only_dinuc_complete}
\alias{only_dinuc_complete}
\title{For generating a position table without pseudo-counts}
\usage{
only_dinuc_complete(table, dinuc_path)
}
\description{
For generating a position table without pseudo-counts
}
\examples{
only_dinuc_complete(table, dinuc_path)


}
