% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments.R
\name{percent_total}
\alias{percent_total}
\title{Calcuate total reads per library and total reads per RNA type}
\usage{
percent_total(x, col_ids)
}
\arguments{
\item{x}{list object of filenames containing depth files}

\item{cold_ids}{column numbers of count data for RNA types to compare to total reads in library}
}
\description{
Calcuate total reads per library and total reads per RNA type
}
\examples{

final_table <- percent_total(count_table, 4:11)

}
