% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_cov_reg}
\alias{plot_cov_reg}
\title{Plot coverage region}
\usage{
plot_cov_reg(data, x, y, cell = NULL, virus = NULL, facet = TRUE,
  start, stop, col_width)
}
\arguments{
\item{data}{table object with count and position data}

\item{x}{variable for x-axis}

\item{y}{variable for y-axis}

\item{cell}{cell}

\item{virus}{virus}

\item{facet}{set to FALSE to generate graphs for a specifc cell and virus combination}

\item{start}{position to start plot region}

\item{stop}{position to end plot region}
}
\description{
Plot coverage region
}
\examples{

plot_cov_reg(viral_cov_small, pos, norm_count, 1800, 2000)

plot_cov_reg(viral_cov_small, pos, norm_count, facet = FALSE, cell = "B6", virus = "ns2", start = 1800, end = 2000, col_width = 0.5)


}
