% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_coverage}
\alias{plot_coverage}
\title{Plot coverage}
\usage{
plot_coverage(data, x, y, cell = NULL, virus = NULL, facet = TRUE)
}
\arguments{
\item{data}{table object with count and position data}

\item{x}{variable for x-axis}

\item{y}{variable for y-axis}

\item{cell}{cell}

\item{virus}{virus}

\item{facet}{set to FALSE to generate graphs for a specifc cell and virus combination}
}
\description{
Plot coverage
}
\examples{

plot_coverage(viral_cov_small, pos, norm_count)

plot_coverage(viral_cov_small, pos, norm_count, facet = FALSE, cell = "B6", virus = "ns2")


}
