% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_depth_single}
\alias{plot_depth_single}
\title{Plot depth for a single variable set}
\usage{
plot_depth_single(data, x, y, cell, virus)
}
\arguments{
\item{data}{table object with count and position data}

\item{x}{variable for x-axis}

\item{y}{variable for y-axis}

\item{cell}{cell}

\item{virus}{virus}
}
\description{
Plot depth for a single variable set
}
\examples{
plot_depth_single(
  viral_cov_tbl,
  start, norm_count,
  B6, ns2
)

}
