% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_dinuc}
\alias{plot_dinuc}
\title{Plot dinucleotides}
\usage{
plot_dinuc(data, x, y, cell = NULL, virus = NULL, facet = TRUE)
}
\arguments{
\item{data}{table with percent dinucleotide data for cell, virus, and time}

\item{facet}{set to FALSE to generate graphs for specifc fold change comparision}
}
\description{
Plot dinucleotides
}
