% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_RNA_capture}
\alias{plot_RNA_capture}
\title{Plot RNA capture by cell}
\usage{
plot_RNA_capture(data, y)
}
\arguments{
\item{data}{table with column data containing percent or total reads by RNA type}

\item{y}{column name for RNA type of interest}
}
\description{
Plot RNA capture by cell
}
\examples{

plot_RNA_capture(table, RNA_col)

}
