% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_top_foldChange}
\alias{plot_top_foldChange}
\title{Plot fold change}
\usage{
plot_top_foldChange(data, x, sample_comp = NULL, facet = TRUE)
}
\arguments{
\item{data}{table with log2 fold change data by cell or by virus}

\item{x}{variable for x-axis, based on a cell or virus fold change analysis}

\item{sample_comp}{variable to filter for speciifc fold change comparisions}

\item{facet}{set to FALSE to generate graphs for specifc fold change comparision}
}
\description{
Plot fold change
}
\examples{

plot_top_foldChange(data, cell)
plot_top_foldChange(data, cell, "MHVS_log2FC", facet = FALSE)


}
