% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{r5.8S_cov_tbl}
\alias{r5.8S_cov_tbl}
\title{5.8S coverage table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3168 rows and 7 columns.}
\usage{
r5.8S_cov_tbl
}
\description{
5.8S coverage table
}
\keyword{datasets}
