% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractive_analysis.R
\name{RNaseL_sub}
\alias{RNaseL_sub}
\title{Normalize count data using the signal that occurs in the abscence of RNaseL}
\usage{
RNaseL_sub(x)
}
\arguments{
\item{x}{viral_cov_table}
}
\description{
Normalize count data using the signal that occurs in the abscence of RNaseL
}
\examples{

RNaseL_sub_tbl <- RNaseL_sub(viral_cov_table)


}
