% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_color_endoU}
\alias{scale_color_endoU}
\title{Color scale constructor for endoU colors}
\usage{
scale_color_endoU(palette = "general", discrete = TRUE,
  reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in endoU_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for endoU colors
}
\examples{
scale_color_endoU("grey")

}
