% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_change_analysis.R
\name{top_sites}
\alias{top_sites}
\title{Idenitfy the top sites of cleavage by fold change analysis}
\usage{
top_sites(df, cols_gather, top)
}
\arguments{
\item{cols_gather}{range of column numbers containing fold change values}

\item{top}{number of top value positions to identify}

\item{data}{fold_change_tbl}
}
\description{
Idenitfy the top sites of cleavage by fold change analysis
}
\examples{

top_sites(fold_change_tbl, cols_gather = 7:8, top = 20)



}
