% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{U6_cov_tbl}
\alias{U6_cov_tbl}
\title{U6 cov table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1476 rows and 6 columns.}
\usage{
U6_cov_tbl
}
\description{
U6 cov table
}
\keyword{datasets}
