% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viral_cov_small}
\alias{viral_cov_small}
\title{Viral coverage table - small version}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10000 rows and 7 columns.}
\usage{
viral_cov_small
}
\description{
Viral coverage table - small version
}
\keyword{datasets}
