% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viral_cov_tbl}
\alias{viral_cov_tbl}
\title{Viral coverage table}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1128852 rows and 7 columns.}
\usage{
viral_cov_tbl
}
\description{
Viral coverage table
}
\keyword{datasets}
