% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{viral_cov_tbl1}
\alias{viral_cov_tbl1}
\title{Viral coverage table -1 +1 posisitions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1102428 rows and 7 columns.}
\usage{
viral_cov_tbl1
}
\description{
Viral coverage table -1 +1 posisitions
}
\keyword{datasets}
