% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqs.R
\name{calc_freq_enrichment}
\alias{calc_freq_enrichment}
\title{Calculate enrichment and significance of dinucleotides captured in the
experimental data compared to prevalence of dinucelotide in RNA sequence of interest}
\usage{
calc_freq_enrichment(observed_freq, experimental_freq)
}
\arguments{
\item{obs_freq}{table contianing dincucleotide frequencies for RNA sequence of interest}

\item{exp_freq}{table containing dinucleotide frequencies calculated from experimental data}
}
\description{
Calculate enrichment and significance of dinucleotides captured in the
experimental data compared to prevalence of dinucelotide in RNA sequence of interest
}
\examples{
combined_freqs <- calc_freqs_enrichment(mhv_freqs, exp_freqs)

}
