% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_depth_multi}
\alias{plot_depth_multi}
\title{Exploratory plots}
\usage{
plot_depth_multi(data, x, y)
}
\arguments{
\item{data}{table object with count and position data}

\item{x}{variable for x-axis}

\item{y}{variable for y-axis}
}
\description{
Exploratory plots
}
