% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtractive_analysis.R
\name{spread_data}
\alias{spread_data}
\title{Normalize count data using the signal that occurs in the abscence of RNaseL or nsp15}
\usage{
spread_data(table, key)
}
\arguments{
\item{key}{variable to spread data based on}

\item{data}{viral_cov_table}
}
\description{
Normalize count data using the signal that occurs in the abscence of RNaseL or nsp15
}
\examples{

Spread the data by cell type or virus type and normalized count

spread_data(viral_cov_tbl, cell)


}
